pyproject.toml
[project]
name = "flask-epl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
]
